feat: entity 생성자 구현
- refactor: 채팅 관련 entity들 entity 디렉토리 하위로 이동
