+++ src/main/java/team03/mopl/domain/chat/entity/ChatRoom.java
@@ -1,42 +1,40 @@
-package team03.mopl.domain.chat;
+package team03.mopl.domain.chat.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import team03.mopl.domain.user.User;
-
-import jakarta.persistence.Id;
 import java.util.UUID;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import team03.mopl.domain.content.Content;
+
 
 @Entity
+@Table(name = "rooms")
 @Getter
-@Table(name="chat_messages")
-public class ChatMessage {
+@NoArgsConstructor
+public class ChatRoom {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
   private UUID id;
 
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @JoinColumn(name="sender_id")
-  private User sender;
-
-  @ManyToOne(fetch = FetchType.LAZY, optional = false)
-  @JoinColumn(name="room_id")
-  private ChatRoom chatRoom;
-
-  @Column(name = "content", nullable = false)
-  private String content;
-
   @CreatedDate
   @Column(name = "created_at", nullable = false, updatable = false)
   private LocalDateTime createdAt;
+
+  @ManyToOne
+  @JoinColumn(name = "content_id", nullable = false)
+  private Content content;
+
+  public ChatRoom(Content content) {
+    this.content = content;
+  }
 }
